tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.clip.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.clip</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.clip</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(async function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

await (function(canvas, ctx) {
ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
ctx.rect(-10, -10, 1, 1);
ctx.clip();
fetch('/images/red.png')
  .then(response => response.blob())
    .then(blob => {
      createImageBitmap(blob)
        .then(bitmap => {
        ctx.fillStyle = '#0f0';
        ctx.fillRect(0, 0, 100, 50);
        ctx.rect(-10, -10, 1, 1);
        ctx.clip();
        ctx.drawImage(document.getElementById('red.png'), 0, 0);
        _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);
    });
});
})(canvas, ctx);
t.done();

});
</script>
